Session replay, Recorded tests, Improved ClassicSession
- Added session recording and replay support to ISession, with concrete
  implementation for classic protocol (no X protocol yet)
- Added a mysqlshrec executable with support for session recording
- Modified mysqlprovision to use mysqlsh as its Python interpreter and
  use the built-in MySQL connection support

- Added query() method with placeholder substitution to ClassicSession
- Added getSession() alias to ClassicSession
- Added DB exception translation to DBError in Python

- Support for executing mysqlprovision with Python 2.6 (bundling
  ConfigParser.py from Python 2.7 distribution)

- Added session traces for 8.0.3 server

+++ mysqlshdk/libs/utils/utils_process.cc
@@ -18,9 +18,12 @@
  */
 
 #include "mysqlshdk/libs/utils/utils_process.h"
+#include <cstdio>
+#include <cstring>
 #ifndef _WIN32
 #include <signal.h>
 #endif
+#include <errno.h>
 #include "mysqlshdk/libs/utils/utils_file.h"
 
 namespace mysqlshdk {
@@ -28,7 +31,12 @@ namespace utils {
 
 #ifndef _WIN32
 bool check_lock_file(const std::string &path) {
-  std::string data = shcore::get_text_file(path);
+  std::string data;
+  if (!shcore::load_text_file(path, data)) {
+    if (errno == ENOENT)
+      return false;
+    throw std::runtime_error(path + ": " + strerror(errno));
+  }
   int64_t pid = std::stoi(data);
   if (kill(pid, 0) < 0) {
     if (errno == ESRCH) {
